fix(redskilled): carry the deaths block into its own module Refs #3499
The two CI failures the /go run left on #3500, and neither is in the
behaviour the issue asked for.

`RedskilledStatuslineDeaths` gained two required fields — the
sender-attributed count and its newest verdict — and the VS Code
extension is a THIRD consumer of that payload with a fixture of its
own. It typechecks against the daemon's interface, so a field the
daemon always states is a field the fixture must state; the canned
verdict names systemd-oomd, so both counts coincide there on purpose.

The file-size ratchet is shrink-only and `statusline-payload.ts` grew
past its 996-line entry, so the answer is decomposition rather than a
raised number. `statusline-deaths.ts` now owns the death block as a
DOMAIN: what counts as a death, which subset a head may print, and
when repetition becomes a boot loop are three judgements with one
owner, and they were living inside a file that was also assembling the
host, the Workers, the budget and the engine.

It imports nothing from the payload builder. `RedskilledDeathObservation`
travelled with it because the observation is a death fact, which keeps
the dependency one-way and spares the module the payload-consumer
exception `statusline-string` would otherwise have had to widen — the
guard caught exactly that and was right to.

Baseline lowered 996 -> 864.
